Octree structure

+++ lib/src/tools.cuh
@@ -11,9 +11,9 @@ using namespace std;
 
 namespace tools {
 
-    unique_ptr<CudaArray<Vector3>> generate_point_cloud_cuboid(unsigned int sideLength);
+    unique_ptr<CudaArray<Vector3>> generate_point_cloud_cuboid(uint64_t sideLength);
     void printKernelDimensions(dim3 block, dim3 grid);
-    void create1DKernel(dim3 &block, dim3 &grid, uint32_t pointCount);
+    void create1DKernel(dim3 &block, dim3 &grid, uint64_t pointCount);
 
     __device__ Vector3 subtract(const Vector3 &a,const Vector3 &b);
     __device__ uint64_t calculateGridIndex(const Vector3 &point, PointCloudMetadata const &metadata, uint16_t gridSize);
